2025_LLM_Agents_course_and_develpoe/Audio_Agent_meeting_recorder/TalkingAI/requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
asyncio
deepgram-sdk
PyAudio
streamlit
beautifulsoup4
sentence-transformers
faiss-cpu
langchainhub